Add django project files and core app
The only changes to the core app other than the default
boilerplate is in the models folder and the resulting
migrations.

There is a single test model with it's migrations.
The model was created with two fields and a unique
contraint on those fields.
A migration was added to add a new field.
Another migration was added to alter the unique
constraint of the model to include the new field.

Migrating the core app fully and then running sqlmigrate
on the 0003 core migration should result in sqlmigrate
failing.